# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Energy Monitor API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/config/service/{hostName}/{serviceDescription}": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "Get all services or single service in Emca Monitor",
                "description": "Retrieves a list of all Services or a specific service by ID or name.",
                "operationId": "getAllServices",
                "parameters": [
                    {
                        "name": "serviceDescription",
                        "in": "path",
                        "description": "ID or name of the service (optional)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "hostName",
                        "in": "path",
                        "description": "ID or name of the host (optional)",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of services",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {}
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "basicAuth": []
                    }
                ]
            }
        },
...
        "/api/config/service/{id}": {
            "delete": {
                "tags": [
                    "Services"
                ],
                "summary": "Delete a service in op5 Monitor",
                "description": "Deletes an existing service.",
                "operationId": "deleteService",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the service to be deleted",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Service deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Service deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Service not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Service not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "basicAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "basicAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "basic"
            }
        }
    }
}